% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFormularFromCube.R
\name{GetFormularFromCube}
\alias{GetFormularFromCube}
\title{Get formular for the summary statistics in RDF data cube}
\usage{
GetFormularFromCube(store, forsparqlprefixcommon)
}
\arguments{
\item{store}{RDF store containing the RDF data cube}

\item{forsparqlprefixcommon}{Prefix for making the query}
}
\value{
expression with the formulars representing the RDF data cube
}
\description{
Get formular for the summary statistics in RDF data cube
}

