% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSQLFromCube.R
\name{GetSQLFromCube}
\alias{GetSQLFromCube}
\title{Get SQL expression for the summary statistics in RDF data cube}
\usage{
GetSQLFromCube(store, dsdName = NULL, srcDsName = "adsl")
}
\arguments{
\item{store}{The name for the rrdf store}

\item{dsdName}{The name of the data frame with data (results)}

\item{srcDsName}{The name of the data frame with underlying data (source)}
}
\value{
SQL expression with the formulars representing the RDF data cube
}
\description{
Get SQL expression for the summary statistics in RDF data cube
}

